feat: :sparkles: metadata workflow update
